toolset/requirements.txt
asciitree==0.3.3
bcrypt==3.1.7
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
dbus-python==1.2.16
defusedxml==0.6.0
distro==1.4.0
docutils==0.16
empy==3.3.2
fasteners==0.19
fonttools==4.55.3
genmsg==0.5.14
genpy==0.6.9
importlib-resources==6.4.5
kiwisolver==1.4.7
-e git+https://e.coding.net/ourbot/mobilemanipulator/manip-dataset-toolset.git@9d6abdfdca29acc1d3ef136b7d7fd79bdc04f2ef#egg=manip_dataset_toolset
matplotlib==3.7.5
netifaces==0.10.4
nose==1.3.7
numcodecs==0.12.1
numpy==1.24.4
packaging==24.2
paramiko==2.6.0
pillow==10.4.0
psutil==6.1.1
pycrypto==2.6.1
pycryptodomex==3.6.1
PyGObject==3.36.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyquaternion==0.9.9
python-dateutil==2.7.3
python-gnupg==0.4.5
PyYAML==5.3.1
roman==2.0.0
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.0
rosdistro-modules==1.0.0
rosinstall==0.7.8
rospkg==1.5.1
rospkg-modules==1.5.1
six==1.14.0
vcstools==0.1.42
wstool==0.1.18
zipp==3.20.2
zarr
scipy
h5py
cv_bridge
tqdm
opencv-python
pyquaternion
